Update and rename Task1.cpp to Task01.cpp
The provided solution was working, but was somewhat confusing with all the different indices! I did a refactor on your code by using the fundamental string manipulation functions, please peruse over the code and ask if you have any questions! Also, when printing strings, **do not iterate** through the whole char array, but just do `std::cout << str << std::endl`, because in strings, we have terminating zero which tells us where the end of the string is (as a sentinel)!