SEM.09: Add 01-02 | Pract.07 Add 05 | Maksim Hristov (#51)
* 'week-4-task15-16'

* 'week4-prac'

* 'prac4'

* 'separate-files'

* 'Sem09:Task1'

* Revert "'Sem09:Task1'"

This reverts commit f34670490cd8c639232b82020a03d40734de1e21.

* 'Sem09:Task1'

* 'Sem09:Task2'

* 'Prac07:Task5'

* 'bug-fix'

* Delete Sem. 04/Pract.04/task17.cpp

* 'task2-enhancement'

* 'separating-logic'

* Update and rename Task1.cpp to Task01.cpp

The provided solution was working, but was somewhat confusing with all the different indices! I did a refactor on your code by using the fundamental string manipulation functions, please peruse over the code and ask if you have any questions! Also, when printing strings, **do not iterate** through the whole char array, but just do `std::cout << str << std::endl`, because in strings, we have terminating zero which tells us where the end of the string is (as a sentinel)!

* Update and rename Task2.cpp to Task02.cpp

Do not make functions for printing strings, just `std::cout << str << std::endl`

* Update and rename Task5.cpp to Task05.cpp

---------

Co-authored-by: Slavi Rezashki <57730123+Slavi15@users.noreply.github.com>